requirements.txt
smart_open==5.0.0
boto3
pandas
pyyaml
absl-py
pyarrow
joblib
dill